Disable the mouse
Mac OS X 10.11 Terminal.app finally supports mice but it does not
support disabling it using a modifier key to e.g. copy a filename from
the status line. Mac OS X < 10.11 has trained me to not depend on the
mouse so the lack of copy & paste is very annoying.

Default to disabled and let me enable it explicitly for the session if I
want to.